Interface_Ethernet: Change LAN9303 VDD18TX1 pin type to input
The TX1 and TX2 pins are according to the data sheet power pins. But the
datasheet also requires the user to connect them together.

Connecting the pins together (as is required by the datasheet) will
cause an ERC error if they are both outputs.
The pins are not supposed to be used to power anything externally
(they are normally connected together and connected to a couple of
capacitors) so having at least one of them as output still makes sense.


See merge request https://gitlab.com/kicad/libraries/kicad-symbols/-/merge_requests/5143