LED: Add more LED_Cree-XHP50 variants and rework existing symbols
Added more symbols for the Cree-XHP50 LED series. There used to be only 2:

* LED_Cree_XHP50_6V
* LED_Cree_XHP50_12V

Now there are 6:

* LED_Cree_XHP50_3V_HighDensity
* LED_Cree_XHP50_6V_HighDensity
* LED_Cree_XHP50_12V_HighDensity
* LED_Cree_XHP50_3V_HighIntensity
* LED_Cree_XHP50_6V_HighIntensity
* LED_Cree_XHP50_12V_HighIntensity

The HighIntensity symbols are the same as the HighDensity symbols, but the footprints and the 3D models are different. I fixed all the KLC check errors for 6V and 12V, changed the description a bit, and also fixed the datasheet URL. The 3V symbol is made from scratch and I made it look like it is in the datasheet on page 44.

See merge request https://gitlab.com/kicad/libraries/kicad-symbols/-/merge_requests/5081