Cleanup style of single h-bridge motor drivers
If possible, the symbols are 600mil wide left to right
The IN1/IN2 pins are placed 100mil above the center line on the left side.
The respective output pins opposite on the right side.

* Review fixes round1

* Fix more KLC warnings in Driver_Motor

PinPositionOffset was 40 for some parts, set to 20
Keywords for the ZXBM5210* had commas

* Update Pin length to 100mil on some motor drivers

They were 200mil which violates KLC

* Cleanup style of single h-bridge motor drivers

If possible, the symbols are 600mil wide left to right
The IN1/IN2 pins are placed 100mil above the center line on the left side.
The respective output pins opposite on the right side.

See merge request https://gitlab.com/kicad/libraries/kicad-symbols/-/merge_requests/5216